+++ example/lib/pages/complex/load_more.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
+import 'package:extended_nested_scroll_view/extended_nested_scroll_view.dart';
 import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
 import 'package:flutter/material.dart';
-import 'package:extended_nested_scroll_view/extended_nested_scroll_view.dart';
 import 'package:loading_more_list/loading_more_list.dart';
 
 @FFRoute(
@@ -21,13 +21,14 @@ class LoadMoreDemo extends StatefulWidget {
 
 class _LoadMoreDemoState extends State<LoadMoreDemo>
     with TickerProviderStateMixin {
-  TabController primaryTC;
+  late final TabController primaryTC;
   final GlobalKey<ExtendedNestedScrollViewState> _key =
       GlobalKey<ExtendedNestedScrollViewState>();
+
   @override
   void initState() {
-    primaryTC = TabController(length: 2, vsync: this);
     super.initState();
+    primaryTC = TabController(length: 2, vsync: this);
   }
 
   @override
@@ -92,8 +93,8 @@ class _LoadMoreDemoState extends State<LoadMoreDemo>
             child: TabBarView(
               controller: primaryTC,
               children: const <Widget>[
-                TabViewItem(),
-                TabViewItem(),
+                TabViewItem(Key('Tab0')),
+                TabViewItem(Key('Tab1')),
               ],
             ),
           )
@@ -117,19 +118,15 @@ class LoadMoreListSource extends LoadingMoreBase<int> {
 }
 
 class TabViewItem extends StatefulWidget {
-  const TabViewItem();
+  const TabViewItem(this.uniqueKey);
+  final Key uniqueKey;
   @override
   _TabViewItemState createState() => _TabViewItemState();
 }
 
 class _TabViewItemState extends State<TabViewItem>
     with AutomaticKeepAliveClientMixin {
-  LoadMoreListSource source;
-  @override
-  void initState() {
-    source = LoadMoreListSource();
-    super.initState();
-  }
+  late final LoadMoreListSource source = LoadMoreListSource();
 
   @override
   void dispose() {
@@ -138,21 +135,25 @@ class _TabViewItemState extends State<TabViewItem>
   }
 
   @override
-  Widget build(BuildContext context) {
-    super.build(context);
-    final LoadingMoreList<int> child = LoadingMoreList<int>(ListConfig<int>(
-        itemBuilder: (BuildContext c, int item, int index) {
-          return Container(
-            alignment: Alignment.center,
-            height: 60.0,
-            child: Text(': ListView$index'),
-          );
-        },
-        sourceList: source));
-
-    return child;
-  }
+  bool get wantKeepAlive => true;
 
   @override
-  bool get wantKeepAlive => true;
+  Widget build(BuildContext context) {
+    super.build(context);
+    final Widget child = ExtendedVisibilityDetector(
+        child: LoadingMoreList<int>(
+          ListConfig<int>(
+            sourceList: source,
+            itemBuilder: (BuildContext c, int item, int index) {
+              return Container(
+                alignment: Alignment.center,
+                height: 60.0,
+                child: Text(': ListView$index'),
+              );
+            },
+          ),
+        ),
+        uniqueKey: widget.uniqueKey);
+    return child;
+  }
 }