+++ lib/src/extended_nested_scroll_view_part.dart
@@ -68,13 +68,37 @@ class _ExtendedNestedScrollCoordinator extends _NestedScrollCoordinator {
       if (actived.isEmpty) {
         for (final _ExtendedNestedScrollPosition scrollPosition
             in _innerController.nestedPositions) {
-          final RenderObject? renderObject =
-              scrollPosition.context.storageContext.findRenderObject();
-          if (renderObject == null || !renderObject.attached) {
+          // TODO(zmtzawqlp): throw exception even mounted is true
+          // In order for an element to have a valid renderObject, it must be '
+          //  'active, which means it is part of the tree.\n'
+          //  'Instead, this element is in the $_lifecycleState state.\n'
+          //  'If you called this method from a State object, consider guarding '
+          //  'it with State.mounted.
+          try {
+            if (!(scrollPosition.context as ScrollableState).mounted) {
+              continue;
+            }
+            final RenderObject? renderObject =
+                scrollPosition.context.storageContext.findRenderObject();
+            if (renderObject == null || !renderObject.attached) {
+              continue;
+            }
+
+            final VisibilityInfo? visibilityInfo =
+                ExtendedVisibilityDetector.of(
+                    scrollPosition.context.storageContext);
+            if (visibilityInfo != null && visibilityInfo.visibleFraction == 1) {
+              if (kDebugMode) {
+                print('${visibilityInfo.key} is visible');
+              }
+              return <_ExtendedNestedScrollPosition>[scrollPosition];
+            }
+
+            if (renderObjectIsVisible(renderObject, bodyScrollDirection)) {
+              return <_ExtendedNestedScrollPosition>[scrollPosition];
+            }
+          } catch (e) {
             continue;
-          }
-          if (renderObjectIsVisible(renderObject, bodyScrollDirection)) {
-            return <_ExtendedNestedScrollPosition>[scrollPosition];
           }
         }
         return _innerController.nestedPositions;
@@ -322,4 +346,21 @@ class _ExtendedRenderSliverFillRemainingWithScrollable
 // I/flutter (14963): -5.684341886080802e-14
 extension DoubleEx on double {
   bool get notZero => abs() > precisionErrorTolerance;
+  bool get isZero => abs() < precisionErrorTolerance;
+}
+
+class _ExtendedNestedInnerBallisticScrollActivity
+    extends _NestedInnerBallisticScrollActivity {
+  _ExtendedNestedInnerBallisticScrollActivity(
+    super.coordinator,
+    super.position,
+    super.simulation,
+    super.vsync,
+    super.shouldIgnorePointer,
+  );
+  @override
+  bool applyMoveTo(double value) {
+    // https://github.com/flutter/flutter/pull/87801
+    return delegate.setPixels(coordinator.nestOffset(value, delegate)).isZero;
+  }
 }